### 2022-05-11 #4385 @arfon
:wave: @webbjj, thanks for your submission to JOSS. Does this submission accompany a related paper in a AAS journal?
