### 2022-05-11 #4385 @webbjj
Hi @arfon , no the submission does not accompany a related paper in a AAS journal.
