+++ tests/test_operations.py
@@ -32,11 +32,11 @@ def test_to_pckms(tol=0.0001,ro=solar_ro,vo=solar_vo):
 
 	tbar,rbar,vbar,zmbar=2.1,5.3,2.5,7553.7
 
-	init_units=['kpckms','galpy','nbody','WDunits']
-	tscales=[1000,(1000.*conversion.time_in_Gyr(ro=ro,vo=vo)),tbar,1000.0]
-	xscales=[1000,ro*1000,rbar,1000.]
-	vscales=[1.,vo,vbar,220.0/conversion.velocity_in_kpcGyr(220.0, 8.0)]
-	mscales=[1.,conversion.mass_in_msol(ro=ro,vo=vo),zmbar,222288.4543021174]
+	init_units=['kpckms','galpy','nbody','WDunits','pcmyr','kpcgyr']
+	tscales=[1000,(1000.*conversion.time_in_Gyr(ro=ro,vo=vo)),tbar,1000.0,1.,1000.]
+	xscales=[1000,ro*1000,rbar,1000.,1.,1000.]
+	vscales=[1.,vo,vbar,220.0/conversion.velocity_in_kpcGyr(220.0, 8.0),1./1.022712165045695,1./1.022712165045695]
+	mscales=[1.,conversion.mass_in_msol(ro=ro,vo=vo),zmbar,222288.4543021174,1.,1.]
 
 	for i in range(0,len(init_units)):
 		t=1.
@@ -60,18 +60,53 @@ def test_to_pckms(tol=0.0001,ro=solar_ro,vo=solar_vo):
 		final=[cluster.tphys,cluster.m,cluster.x,cluster.y,cluster.z,cluster.vx,cluster.vy,cluster.vz,cluster.xc,cluster.yc,cluster.zc,cluster.vxc,cluster.vyc,cluster.vzc,cluster.xgc,cluster.ygc,cluster.zgc,cluster.vxgc,cluster.vygc,cluster.vzgc]
 		scale_test(init,final,tscales[i],mscales[i],xscales[i],vscales[i])
 
+def test_to_pcmyr(tol=0.0001,ro=solar_ro,vo=solar_vo):
+
+	tbar,rbar,vbar,zmbar=2.1,5.3,2.5,7553.7
+
+	init_units=['kpckms','galpy','nbody','WDunits','pckms','kpcgyr']
+	tscales=[1000,(1000.*conversion.time_in_Gyr(ro=ro,vo=vo)),tbar,1000.0,1.,1000.]
+	xscales=[1000,ro*1000,rbar,1000.,1.,1000.]
+	vscales=[1.022712165045695,vo*1.022712165045695,vbar*1.022712165045695,1.,1.022712165045695,1.]
+	mscales=[1.,conversion.mass_in_msol(ro=ro,vo=vo),zmbar,222288.4543021174,1.,1.]
+
+	for i in range(0,len(init_units)):
+		print('DEBUG: ',i,init_units[i])
+		t=1.
+		m=np.random.rand(100)
+		x,y,z=np.random.rand(100),np.random.rand(100),np.random.rand(100)
+		vx,vy,vz=np.random.rand(100),np.random.rand(100),np.random.rand(100)
+		xc,yc,zc,vxc,vyc,vzc=np.random.rand(6)
+		xgc,ygc,zgc,vxgc,vygc,vzgc=np.random.rand(6)
+
+		cluster=ctools.StarCluster(tphys=t,units=init_units[i],origin='cluster')
+		cluster.add_stars(x,y,z,vx,vy,vz,m=m,analyze=True)
[…]
