### 2022-06-02 d78f9ef
set potential to nan when two stars have r=0
