### 2022-06-02 4a4d17c
set r to 1e-10 when calculating energies at r=0
