### 2022-06-02 0b09f88
set r to nan when calculating energies at r=0
