### 2022-06-03 #4385 @christinahedges
Hi @webbjj, thanks for your submission to JOSS! I'm going to start by reading through your submission, and then I will ping some potential reviewers. We will find at least two reviewers, and then start the review. They will go through a checklist and then provide you feedback on `clustertools`. Here is one point that I've noted as I looked through your submission:

1. I see this tool is largely for using the output of other N-body simulations. This can absolutely be a valuable, substantial contribution, but in these instances we want to be sure. Given my check of your submission, I'm happy to have this reviewed. I will encourage some discussion from the reviewers to check that they believe this is a substantial and useful contribution to the community.

I have some potential reviewers in mind that I can start with, but sometimes it can be hard to find reviewers who have no [conflicts of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). If you know of researchers who have relevant expertise and are not conflicted with you, please feel free to suggest their names here, otherwise I will search for reviewers and get back to you.

It can sometimes take a while to get together reviewers, but after we find them the review is usually complete in 4-6 weeks.
