+++ clustertools/util/coordinates.py
@@ -167,7 +167,7 @@ def cyl_coords(cluster):
 
     R,phi,z,vR,vT,vz=cart_to_cyl(cluster.x,cluster.y,cluster.z,cluster.vx,cluster.vy,cluster.vz)
 
-    cluster.return_cluster(units0,origin0, rorder0, rorder_origin0)
+    cluster.return_cluster(units0, origin0, rorder0, rorder_origin0)
 
     return R,phi,z,vR,vT,vz
 
