### 2022-06-16 #4483 @christinahedges
**FYI for both reviewers:** In the pre-review we discussed with the submitting authors that this tool is designed to improve the interface with products from several other tools. We believe this type of tool can still count under JOSS's requirement for [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) and can meet the JOSS requirements including [having an obvious research application](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) and utility to the community. I want to bring your attention to this during your review, to make sure that I can check in with both reviewers that they agree this tool meets those standards.
