### 2022-06-17 #4483 @rieder
A question:
The review specifies version 1.0 of Clustertools. However, it seems that this version was released two years ago (according to the GitHub tag). I guess we should look at the current version instead? Maybe a new tag would be appropriate?
