### 2022-07-06 d127081
debug dtout call in gyrfalcon load
