+++ clustertools/io/load.py
@@ -222,6 +222,9 @@ def load_cluster(
         # Read in snapshot from gyrfalcon.
         filename=_get_filename(filename,**kwargs)
 
+        deltat=kwargs.pop('deltat',1)
+        dtout=kwargs.pop('dtout',deltat)
+
         if dtout is None:
             filein = open(filename, "r")
             filein.readlines()
