### 2022-07-06 32ca1f5
debug dtout call in gyrfalcon load
