### 2022-07-06 74320b6
debug dtout call in gyrfalcon load
