### 2022-07-06 8db4107
debug dtout call in gyrfalcon load
