### 2022-07-11 40f17ba
internal and external call of energies now returns values for individual stars
