### 2022-07-12 9cea0ed
Added support for multiple indeces in energies function and method
