### 2022-07-13 636466d
clean variables for kinetic energy for internal and external energies call
