### 2022-07-13 b4fa3a0
fix docstring on potential energy functions
