### 2022-07-13 e29493e
Added support for mutiple indeces in energy functions
