### 2022-07-14 e002fa0
Combined ids and i_d into one argument for the energies function and method
