### 2022-07-14 6f54c09
Combined ids and i_d into one argument for the energies function and method
