+++ clustertools/cluster/cluster.py
@@ -2169,8 +2169,9 @@ class StarCluster(object):
         Returns
         -------
         kin,pot : float
-          kinetic and potential energy of every star
-
+          kinetic and potential energy of every star if the i_d argument is not used. If i_d
+          argument is used, return an arrays with potential and kinetic energy in the same shape
+          of i_d
         History
         -------
            2019 - Written - Webb (UofT)
