### 2022-07-14 054200a
Combined ids and i_d into one argument for the energies function and method
