### 2022-07-14 527ad71
Combined ids and i_d into one argument for the energies function and method
