### 2022-07-14 ad4f91d
Combined ids and i_d into one argument for the energies function and method
