### 2022-07-14 #4483 @webbjj
Thanks @lwang-astro .

I haven't tested the code with anything less than python 3, so I am wondering if that is the issue with the install error. Just to confirm, you normally have permission to install in the directory '/usr/local/lib/python2.7/dist-packages/test-easy-install-118055.write-test'? If so I will try and reproduce the error and track down the problem. It will be helpful to know if it works with python3, but I understand if you are unable to test that.

And thanks for the heads up on the docs, fixing now.
