+++ docs/source/reference/clustertools.rst
@@ -14,7 +14,6 @@ Subpackages
 
    clustertools.analysis
    clustertools.cluster
-   clustertools.custom
    clustertools.io
    clustertools.tidaltail
    clustertools.util
