### 2022-07-14 #4483 @rieder
I would suggest not supporting any Python versions below 3.5 or so. Those older versions are phased out / unsupported anyway, and it avoids a *lot* of headaches.
