+++ setup.py
@@ -13,7 +13,7 @@ setuptools.setup(
     long_description_content_type='text/markdown',
     license='MIT',
     packages=["clustertools","clustertools/analysis","clustertools/cluster","clustertools/io","clustertools/io/data","clustertools/tidaltail","clustertools/util"],
-    setup_requires=['scipy'],
+    setup_requires=['numpy>1.8','scipy'],
     install_requires=['galpy','seaborn','numba>=0.55.2'],
     include_package_data=True,
     )
