### 2022-07-14 132f087
fixing numba/llvmlite in docs
