### 2022-07-14 3e2358e
Combined ids and i_d into one argument for the energies function and method
