### 2022-07-15 #4483 @lwang-astro
@webbjj This installation error is independent of Python vesion, but relate to the default path to install, the same error occur for Python3.8.
```
[lwang@Thinkpad-x1:~/code/clustertools]$ python3 setup.py install
running install
error: can't create or remove files in install directory

The following error occurred while trying to add or remove files in the
installation directory:

    [Errno 13] Permission denied: '/usr/local/lib/python3.8/dist-packages/test-easy-install-231.write-test'

The installation directory you specified (via --install-dir, --prefix, or
the distutils default setting) was:

    /usr/local/lib/python3.8/dist-packages/

Perhaps your account does not have write access to this directory?  If the
installation directory is a system-owned directory, you may need to sign in
as the administrator or "root" account.  If you do not have administrative
access to this machine, you may wish to choose a different installation
directory, preferably one that is listed in your PYTHONPATH environment
variable.

For information on other options, you may wish to consult the
documentation at:

  https://setuptools.readthedocs.io/en/latest/easy_install.html

Please make the appropriate changes for your system and try again.
```

For Linux system (Ubuntu), standard users don't have permission to install package in '/usr/local/lib/'
It only works when 'sudo' is used:
`sudo python3 setup.py install`

I think that this is a general feature for Python installation for either setup.py or pip3, not a problem from clustertools.
For pip3, I can successfully install the package if `--user` option is provided (`pip3 install --user clustertools`). Otherwise, the same issue occurs.
Perhaps this issue does not occur when the package is installed via anaconda or python virtual environment.
I suggest to include some sentences in the manual that 'sudo' is needed to avoid permission error or how to manually set the install path.
