### 2022-07-15 c609720
fixed mfrac in rcore to be 10% for both internal and external calls, doc installation update
