+++ docs/source/installation.rst
@@ -10,12 +10,27 @@ To install ``clustertools``, simply clone the GitHub repository and install via 
 >>> cd clustertools
 >>> python setup.py install
 
+Please note that if you don't have permission to write files to the default install location (often /usr/local/lib), you will either need to run
+
+>>> sudo python setup.py install
+
+or
+
+>>> python setup.py install --prefix='PATH'
+
+where 'PATH' is a directory that you do have permission to write in and is in your PYTHONPATH.
+
 It has recently become possible to install ``clustertools`` using pip:
 
 >>> pip install clustertools
 
-however please note this version is not updated as frequently as the GitHub repository.
+however please note this version is not updated as frequently as the GitHub repository. Similarly, if permissions are an issue, you can use:
 
+>>> pip install --user clustertools
+
+or
+
+>>> pip install clustertools --install-option="--prefix=PATH"
 
 Requirements
 ------------
