### 2022-07-15 #4483 @webbjj
@rieder - thanks for the heads up on python 2.7. I looked at it quickly and seems like a headache and not worth it. I image going from 2.7 to 3 was a lot of work for AMUSE.

@lwang-astro
- I have updated the installation instructions for users that might not have permissions to install in certain directories. I agree it was bad form to assume a user knows how to deal with such issues.
- I have unfortunately not been able to reproduce the issue with either the virial_radius or rcore functions. Did you use pip install or git clone? If the former then perhaps I need to update the pip. The pcmyr have been added recently as I get ready to add PeTar support, so there is a chance there is a bug that my tests have missed as it has not been as heavily used as pckms. Alternatively maybe you could send the input file?
