### 2022-07-15 a533a15
add internal length conversion function
