+++ clustertools/util/__init__.py
@@ -1,4 +1,5 @@
 from . import constants
+from . import units
 from . import coordinates
 from . import output
 from . import plots

[…]
