### 2022-07-15 7fbc885
allow for subset of energies to be calculated
