### 2022-07-15 9b5916e
allow for subset of energies to be calculated
