+++ docs/source/index.rst
@@ -135,6 +135,7 @@ Contributers
 -----------------
 Jo Bovy - https://github.com/jobovy
 Nathaniel Starkman - https://github.com/nstarman
+Erik Gillis - https://github.com/e-gillis
 
 Library Reference
 -----------------
