### 2022-07-15 cf73110
allow for subset of energies to be calculated
