### 2022-07-21 #4483 @lwang-astro
@webbjj
Thank for the update. The new pip version have included the rcore function. But I notice two new issues.

1. The value return from `rcore` is different from the definition used in NBODY6 and PeTar codes.

I used mcluster code (https://github.com/lwang-astro/mcluster) to generate a Plummer model with
rvir = 1.20395   rh = 1.00000    rplummer = 0.76628.
The corresponding commander: `mcluster -N 1000 -P 0 -R 1 -C 5 -u 1 -o N1k >mc.log`.

Using the input data, the returned value from clustertools:

```
cluster.rcore()
0.7325688314176657

cluster.virial_radius()
1.202307596486262
```
It seems that the rcore value is close to rplummer (a), which is the scale radius of Plummer potential.
This is different from the usually assumed core radius, such as the Casertano & Hut (1985) method with the squares of densities used in NBODY6 (also in PeTar). The corresponding value for this example is approximately 0.4465.

2. It seems that if the model include a large fraction of binaries, the virial radius is not properly calculated.
I used mcluster again to generate a Plummer model with N=1000, Rh=1 and 100% binary (see mcluster manual for the detail of primordial binary setup; the corresponding commander: `mcluster -N 1000 -P 0 -R 1 -b 1 -C 5 -u 1 -o N1kb >mc.log`.).
Then the clustertools return a very small virial radius:
```
cluster.virial_radius()
4.6897395944288934e-05
```
I think the c.m. velocities of binaries are needed to used to calculate velocity dispersion. Is any way in clustertools to set binary population like `add_stars`? It seems that the number of binary is only available in the arguments of `add_nbody6` function. Is any general way to include binaries?
