### 2022-07-22 33bc541
new core radius definition and additional binary support
