+++ clustertools/cluster/cluster.py
@@ -175,6 +175,22 @@ class StarCluster(object):
         self.vz = np.array([])
         self.m0 = np.array([])
 
+        # variables for add_binary_stars
+        self.mb1 = np.array([])
+        self.xb1 = np.array([])
+        self.yb1 = np.array([])
+        self.zb1 = np.array([])
+        self.vxb1 = np.array([])
+        self.vyb1 = np.array([])
+        self.vzb1 = np.array([])
+        self.mb2 = np.array([])
+        self.xb2 = np.array([])
+        self.yb2 = np.array([])
+        self.zb2 = np.array([])
+        self.vxb2 = np.array([])
+        self.vyb2 = np.array([])
+        self.vzb2 = np.array([])
+
         #If using Nbodypp
         self.rhos=np.array([])
 
@@ -303,6 +319,8 @@ class StarCluster(object):
         self.semi = np.array([])
         self.m1 = np.array([])
         self.m2 = np.array([])
+        self.m01 = np.array([])
+        self.m02 = np.array([])
         self.logl1 = np.array([])
         self.logl2 = np.array([])
         self.logr1 = np.array([])
@@ -311,6 +329,8 @@ class StarCluster(object):
         self.ep2 = np.array([])
         self.ospin1 = np.array([])
         self.ospin2 = np.array([])
+        self.npop1 = np.array([])
+        self.npop2 = np.array([])
 
         # variables of energies
         self.kin = np.array([])
@@ -363,7 +383,7 @@ class StarCluster(object):
         self.ngroup=0
 
     def add_stars(
-        self, x, y, z, vx, vy, vz,m=None,id=None,m0=None,npop=None,sortstars=False,analyze=True
+        self, x, y, z, vx, vy, vz,m=None,id=None,m0=None,npop=None,nb=0,sortstars=False,analyze=True
     ):
         """Add stars to StarCluster.
 
@@ -383,6 +403,8 @@ class StarCluster(object):
             initial stellar mass
         npop : int
             population number, for use with multiple populations
+        nb: int
+            number of binary stars, which are assumed to be the first 2*nb stars in the array (default:0)
         sortstars: bool
             order stars by radius (default: False)
         analyze : bool
@@ -396,40 +418,122 @@ class StarCluster(object):
 
         """
 
-        self.x = np.append(self.x, np.array(x))
-        self.y = np.append(self.y, np.array(y))
-        self.z = np.append(self.z, np.array(z))
-        self.vx = np.append(self.vx, np.array(vx))
-        self.vy = np.append(self.vy, np.array(vy))
-        self.vz = np.append(self.vz, np.array(vz))
+        #Check for float entries
+        params=([x,y,z,vx,vy,vz])
+        nfloat=np.zeros(len(params),dtype=bool)
+        npmax=0
+        for i,p in enumerate(params):
+            if isinstance(p,float):
+                nfloat[i]=True
+            else:
+                npmax=int(np.maximum(npmax,len(p)))
+
+        if nfloat[0]: x=np.ones(npmax)*x
+        if nfloat[1]: y=np.ones(npmax)*y
+        if nfloat[2]: z=np.ones(npmax)*z
+        if nfloat[3]: vx=np.ones(npmax)*vx
+        if nfloat[4]: vy=np.ones(npmax)*vy
+        if nfloat[5]: vz=np.ones(npmax)*vz
+
+        #Check for bianries
+        if nb>0:
+            if nb==1:
+                arg1,arg2=0,1
+            else:
+                arg1=np.arange(0,2*nb-1,2)
+                arg2=arg1+1
+
+            if m is None:
[…]

[…]
