### 2022-07-22 #4483 @webbjj
Thanks for this check @lwang-astro

- I have added two ways of adding binary stars as outlined in the docs (https://clustertools.readthedocs.io/en/latest/cluster.html#the-starcluster-class). In short, you can now pass nb when adding stars. So in your test you would pass nb=500 when using add_stars. There is also now an add_binary_stars option where you can add the pairs individually. This actually existed in the background when loading Nbody6++ stars. These methods ensure only the centre of mass properties are in the main arrays (cluster.x, cluster.vx, etc.) which seems to fix the virial radius calculation issue you ran into.
- My calculation of the core radius was more observer oriented, and I would look to see where the density drops to 1/3 the central value (1/2 for projection), however I think it would be a good idea to update this to be in line with simulations. So I have implemented the  Casertano & Hut (1985) method, which is now the default. I have also implemented their method for finding the centre of density as an option, but the Harfst approach is still the default.

The fixes have been pushed to github and pip

Two questions for you @lwang-astro:
- Your Petar/tools code for the Casertano & Hut (1985) method obviously served as inspiration for what I have done. Is it OK to put the main PeTar paper as a suggested reference for anyone using the rcore command?
- In PeTar/tools where you using the Casertano & Hut (1985) method you find the mass within r6 as nb_mass_tot6=np.sum(particle[nb_index_list6].mass,axis=1) + particle.mass. I am curious if there is a bug here, or perhaps I misunderstanding either KDTree or the method, because to me it looks like the central star mass is being counted twice since   nb_index_list6 should include the central star. Furthermore, I think the Casertano & Hut (1985) method ignores the outermost star in the mass calculation. So shouldn't it be np.sum(particle[nb_index_list6].mass,axis=1) - particle.mass[nb_index_list6[:,-1]]. That is how I have implemented it in clustertools and it seems to calculate the expected core radius of the Plummer sphere you generate (albeit with 10K particles instead of 1K particles). Interested in your thoughts.
