### 2022-08-05 6057662
gyrfalcon speedup testing
