### 2022-08-08 b9e3a0e
debug gyrfalcon speedup
