### 2022-08-08 f67034f
debug gyrfalcon speedup
