### 2022-08-08 b10b263
debug gyrfalcon speedup
