### 2022-08-09 #4483 @lwang-astro
@webbjj Now the add_star with binaries work fine.
> Your Petar/tools code for the Casertano & Hut (1985) method obviously served as inspiration for what I have done. Is it OK to put the main PeTar paper as a suggested reference for anyone using the rcore command?

Thank you. no problem.

> In PeTar/tools where you using the Casertano & Hut (1985) method you find the mass within r6 as nb_mass_tot6=np.sum(particle[nb_index_list6].mass,axis=1) + particle.mass. I am curious if there is a bug here, or perhaps I misunderstanding either KDTree or the method, because to me it looks like the central star mass is being counted twice since nb_index_list6 should include the central star. Furthermore, I think the Casertano & Hut (1985) method ignores the outermost star in the mass calculation. So shouldn't it be np.sum(particle[nb_index_list6].mass,axis=1) - particle.mass[nb_index_list6[:,-1]]. That is how I have implemented it in clustertools and it seems to calculate the expected core radius of the Plummer sphere you generate (albeit with 10K particles instead of 1K particles). Interested in your thoughts.

Thank you for the check. Indeed here is a bug in petar tool. I have a misunderstood of KDTree.query before. The corrected version of petar tool should be:
```
 76        # 6 nearest neighbors
 77        nb_r_list6, nb_index_list6 = kdtree.query(particle.pos,k=7)
 78        nb_mass_tot6=np.sum(particle.mass[nb_index_list6[:,:-1]],axis=1)
 79
 80        nb_inv_r6 = 1/nb_r_list6[:,-1]
 81        rho = nb_mass_tot6*(nb_inv_r6*nb_inv_r6*nb_inv_r6)

106        rho2 = particle.density*particle.density
107        rc = np.sqrt((particle.r2*rho2).sum()/(rho2.sum()))
```
The new version use k=7 (6 neighbors + itself)
Then the summation remove last one as you suggested.
The radius need to use the outmost member.

After the correction, I found the core radius from PeTar tools seems to be consistent with mcluster result.
I update the pip version of clustertool and found that the cluster.rcore(method='casertano') is 0.52 for a sample of N=1000 from mcluster with no binary.
For comparison, petar tool returns 0.4532 and mcluster returns 0.452780. In mcluster and petar tool, the square weighted version is used, which is different from the original Casertano & Hut method. I guess this may cause the difference of results from clustertool and petar tool.
