### 2022-08-09 #4483 @webbjj
@lwang-astro - interesting. My students and I all use clustertools in jupyter and I have yet to see this error. It might be a version issue, in which case i should add that to the setup file.

Are you able to make other plots in jupyter with latex symbols in the axis labels?
