### 2022-08-09 #4483 @lwang-astro
@webbjj I have no problem to use latex format for axe label. such as ```axe.set_label(r'\sigma')``` works for me.
