### 2022-08-09 #4483 @webbjj
Hmmm, well that is interesting. So on my end I am using:

matplotlib version 3.4.2
jupyter lab version 3.4.3

and I can run everything in clustertools/docs/source/notebooks/profiles.ipynb.

I am not sure where matplotlib looks for the latex plug in. I made sure the latex module was up to date though. I have also found that the galpy.util sometimes messes with my plotting. What version of galpy do you have? I have 1.8.1.dev0.

How does this compare to your setup?
