### 2022-08-10 7ddc702
only include binary COM in NBODY6++ loads
