### 2022-08-10 a854e34
only include binary COM in NBODY6++ loads
