### 2022-08-10 37d4f20
only include binary COM in NBODY6++ loads
