### 2022-08-10 ca041e0
only include binary COM in NBODY6++ loads
