### 2022-08-10 0f249ca
only include binary COM in NBODY6++ loads
