### 2022-08-10 19bb78d
only include binary COM in NBODY6++ loads
