### 2022-08-10 4801fd6
only include binary COM in NBODY6++ loads
