+++ clustertools/io/nbody6pp.py
@@ -157,7 +157,6 @@ def _get_nbody6pp(conf3, bev82=None, sev83=None, snap40=None, ofile=None, advanc
     else:
         ntot,alist,x,y,z,vx,vy,vz,m,i_d,rhos,xns,pot=_get_nbody6pp_conf3(conf3,nsnap=nsnap,**kwargs)
 
-
         cluster = StarCluster(
             alist[0],
             units="nbody",
@@ -179,42 +178,43 @@ def _get_nbody6pp(conf3, bev82=None, sev83=None, snap40=None, ofile=None, advanc
                 binargs=np.arange(0,nb,1)
                 binargs1=np.arange(0,nb,2)
                 binargs2=binargs1+1
-                ssargs=np.arange(2*nb,ntot,1)
             else:
-                ssargs=np.arange(0,ntot,1)
 
             cluster.rhos=np.zeros(ntot-nb)
             pots=np.zeros(ntot-nb)
 
             if nb>0: cluster.rhos[binargs]=rhos[binargs1]+rhos[binargs2]
-            cluster.rhos[nb:]=rhos[ssargs]
+            cluster.rhos[nb:]=rhos[2*nb:]
             pots[binargs]=pot[binargs1]+pot[binargs2]
-            pots[nb:]=pot[ssargs]
+            pots[nb:]=pot[2*nb:]
 
             v=np.sqrt(cluster.vx**2.+cluster.vy**2.+cluster.vz**2.)
             ek=0.5*cluster.m*v**2.
             cluster.add_energies(ek,pots)
 
         if bev82 is not None and sev83 is not None:
+
+            kws=np.ones(ntot-nb)*-10.0
+            zl1s=np.ones(ntot-nb)*-10.0
+            r1s=np.ones(ntot-nb)*-10.0
+
             arg,i_d,kw,ri,m1,zl1,r1,te,i_d1,i_d2,kw1,kw2,kwb,rib,ecc,pb,semi,m1b,m2b,zl1b,zl2b,r1b,r2b,te1,te2=_get_nbody6pp_ev(bev82,sev83,nsnap=nsnap,**kwargs)
             #Convert from fortran array address to python
-            arg-=1
-
-            kws=np.ones(len(x))*-10.0
-            zl1s=np.ones(len(x))*-10.0
-            r1s=np.ones(len(x))*-10.0
 
             if nb>0:
                 kws[binargs]=kwb
                 zl1s[binargs]=zl1b+zl2b
                 r1s[binargs]=r1b+r2b
+                args=arg[2*nb:]-nb-1
+            else:
+                args=arg-1
 
-            kws[nb:]=kw[ssargs]
-            zl1s[nb:]=zl1[ssargs]
-            r1s[nb:]=r1[ssargs]
+            args=args.astype(int)
 
-            if len(kw) != len(x):
-                if verbose: print('SSE/BSE NBODY6++ ERROR',len(x)-len(kw))
+            kws[args]=kw[2*nb:]
+            zl1s[args]=zl1[2*nb:]
+            r1s[args]=r1[2*nb:]
+
             cluster.add_sse(kws,zl1s,r1s)
 
             pb=(10.0**pb)/cluster.tbar_days
@@ -223,6 +223,10 @@ def _get_nbody6pp(conf3, bev82=None, sev83=None, snap40=None, ofile=None, advanc
             m2b/=cluster.zmbar
 
             cluster.add_bse(i_d1,i_d2,kw1,kw2,kwb,ecc,pb,semi,m1b,m2b,zl1b,zl2b,r1b,r2b)
+
+
+    if np.sum(cluster.kw<0) >0:
[…]
