### 2022-08-10 7103aee
only include binary COM in NBODY6++ loads
