### 2022-08-10 3466016
only include binary COM in NBODY6++ loads
