### 2022-08-10 d832c31
debug binaries with nbody6pp
