### 2022-08-13 #4483 @lwang-astro
@webbjj  I have tried the profile.ipynb and the problem occur again. Here I show the full error message. It suggests that
RuntimeError: latex was not able to process the following string:
b'lp'
```
CalledProcessError                        Traceback (most recent call last)
File ~/.local/lib/python3.8/site-packages/matplotlib/texmanager.py:233, in TexManager._run_checked_subprocess(self, command, tex, cwd)
    232 try:
--> 233     report = subprocess.check_output(
    234         command, cwd=cwd if cwd is not None else self.texcache,
    235         stderr=subprocess.STDOUT)
    236 except FileNotFoundError as exc:

File /usr/lib/python3.8/subprocess.py:415, in check_output(timeout, *popenargs, **kwargs)
    413     kwargs['input'] = empty
--> 415 return run(*popenargs, stdout=PIPE, timeout=timeout, check=True,
    416            **kwargs).stdout

File /usr/lib/python3.8/subprocess.py:516, in run(input, capture_output, timeout, check, *popenargs, **kwargs)
    515     if check and retcode:
--> 516         raise CalledProcessError(retcode, process.args,
    517                                  output=stdout, stderr=stderr)
    518 return CompletedProcess(process.args, retcode, stdout, stderr)

CalledProcessError: Command '['latex', '-interaction=nonstopmode', '--halt-on-error', '../aba7304bf2416c0194a8a27472595813.tex']' returned non-zero exit status 1.

RuntimeError                              Traceback (most recent call last)
Input In [8], in <module>
----> 1 rprof, pprof, nprof=cts.rho_prof(cluster,plot=True)

File ~/.local/lib/python3.8/site-packages/clustertools/analysis/profiles.py:222, in rho_prof(cluster, mmin, mmax, rmin, rmax, nrad, vmin, vmax, emin, emax, kwmin, kwmax, npop, indx, bins, projected, normalize, plot, **kwargs)
    219 if normalize:
    220     x/=cluster.rm
--> 222 _lplot(
    223     x,
    224     y,
    225     xlabel=xlabel,
    226     ylabel=ylabel,
    227     title="Time = %f" % cluster.tphys,
    228     log=kwargs.pop('log',True),
    229     overplot=overplot,
    230     filename=filename,
    231     **kwargs,
    232 )
    234 if filename != None:
    235     plt.savefig(filename)

File ~/.local/lib/python3.8/site-packages/clustertools/util/plots.py:312, in _lplot(x, y, ltype, xlabel, ylabel, legend, title, xlim, ylim, scale, log, logx, logy, filename, overplot, **kwargs)
    247 def _lplot(
    248     x,
    249     y,
   (...)
    263     **kwargs
    264 ):
    265     """ Wrapper for plotting lines with matplotlib.pyplot.plot
    266       - allows for most pyplot commands to be assigned when calling the function
    267
   (...)
    310     2018
    311     """
--> 312     return _plot(
    313         x,
    314         y,
    315         ptype=ltype,
    316         xlabel=xlabel,
    317         ylabel=ylabel,
    318         legend=legend,
    319         title=title,
    320         xlim=xlim,
    321         ylim=ylim,
    322         scale=scale,
    323         log=log,
    324         logx=logx,
    325         logy=logy,
    326         filename=filename,
    327         overplot=overplot,
    328         **kwargs
    329     )

File ~/.local/lib/python3.8/site-packages/clustertools/util/plots.py:239, in _plot(x, y, ptype, xlabel, ylabel, legend, title, xlim, ylim, scale, log, logx, logy, filename, overplot, **kwargs)
    236 if legend:
    237     plt.legend()
--> 239 plt.tight_layout()
    241 if filename != None:
    242     plt.savefig(filename)

File ~/.local/lib/python3.8/site-packages/matplotlib/pyplot.py:2302, in tight_layout(pad, h_pad, w_pad, rect)
   2300 @_copy_docstring_and_deprecators(Figure.tight_layout)
   2301 def tight_layout(*, pad=1.08, h_pad=None, w_pad=None, rect=None):
-> 2302     return gcf().tight_layout(pad=pad, h_pad=h_pad, w_pad=w_pad, rect=rect)

File ~/.local/lib/python3.8/site-packages/matplotlib/figure.py:3197, in Figure.tight_layout(self, pad, h_pad, w_pad, rect)
   3195 renderer = _get_renderer(self)
   3196 with getattr(renderer, "_draw_disabled", nullcontext)():
-> 3197     kwargs = get_tight_layout_figure(
   3198         self, self.axes, subplotspec_list, renderer,
   3199         pad=pad, h_pad=h_pad, w_pad=w_pad, rect=rect)
   3200 if kwargs:
   3201     self.subplots_adjust(**kwargs)

File ~/.local/lib/python3.8/site-packages/matplotlib/tight_layout.py:320, in get_tight_layout_figure(fig, axes_list, subplotspec_list, renderer, pad, h_pad, w_pad, rect)
    315         return {}
    316     span_pairs.append((
    317         slice(ss.rowspan.start * div_row, ss.rowspan.stop * div_row),
    318         slice(ss.colspan.start * div_col, ss.colspan.stop * div_col)))
--> 320 kwargs = _auto_adjust_subplotpars(fig, renderer,
    321                                   shape=(max_nrows, max_ncols),
    322                                   span_pairs=span_pairs,
    323                                   subplot_list=subplot_list,
    324                                   ax_bbox_list=ax_bbox_list,
    325                                   pad=pad, h_pad=h_pad, w_pad=w_pad)
    327 # kwargs can be none if tight_layout fails...
    328 if rect is not None and kwargs is not None:
    329     # if rect is given, the whole subplots area (including
    330     # labels) will fit into the rect instead of the
   (...)
    334     # auto_adjust_subplotpars twice, where the second run
    335     # with adjusted rect parameters.

File ~/.local/lib/python3.8/site-packages/matplotlib/tight_layout.py:82, in _auto_adjust_subplotpars(fig, renderer, shape, span_pairs, subplot_list, ax_bbox_list, pad, h_pad, w_pad, rect)
     80 if ax.get_visible():
     81     try:
---> 82         bb += [ax.get_tightbbox(renderer, for_layout_only=True)]
     83     except TypeError:
     84         bb += [ax.get_tightbbox(renderer)]

File ~/.local/lib/python3.8/site-packages/matplotlib/axes/_base.py:4619, in _AxesBase.get_tightbbox(self, renderer, call_axes_locator, bbox_extra_artists, for_layout_only)
   4617 if self.xaxis.get_visible():
   4618     try:
-> 4619         bb_xaxis = self.xaxis.get_tightbbox(
   4620             renderer, for_layout_only=for_layout_only)
   4621     except TypeError:
   4622         # in case downstream library has redefined axis:
   4623         bb_xaxis = self.xaxis.get_tightbbox(renderer)

File ~/.local/lib/python3.8/site-packages/matplotlib/axis.py:1105, in Axis.get_tightbbox(self, renderer, for_layout_only)
   1101     return
   1103 ticks_to_draw = self._update_ticks()
-> 1105 self._update_label_position(renderer)
   1107 # go back to just this axis's tick labels
   1108 ticklabelBoxes, ticklabelBoxes2 = self._get_tick_bboxes(
   1109             ticks_to_draw, renderer)

File ~/.local/lib/python3.8/site-packages/matplotlib/axis.py:2083, in XAxis._update_label_position(self, renderer)
   2079     return
   2081 # get bounding boxes for this axis and any siblings
   2082 # that have been set by `fig.align_xlabels()`
-> 2083 bboxes, bboxes2 = self._get_tick_boxes_siblings(renderer=renderer)
   2085 x, y = self.label.get_position()
   2086 if self.label_position == 'bottom':

File ~/.local/lib/python3.8/site-packages/matplotlib/axis.py:1881, in Axis._get_tick_boxes_siblings(self, renderer)
   1879 axis = getattr(ax, f"{axis_name}axis")
   1880 ticks_to_draw = axis._update_ticks()
-> 1881 tlb, tlb2 = axis._get_tick_bboxes(ticks_to_draw, renderer)
   1882 bboxes.extend(tlb)
   1883 bboxes2.extend(tlb2)

File ~/.local/lib/python3.8/site-packages/matplotlib/axis.py:1085, in Axis._get_tick_bboxes(self, ticks, renderer)
   1083 def _get_tick_bboxes(self, ticks, renderer):
   1084     """Return lists of bboxes for ticks' label1's and label2's."""
-> 1085     return ([tick.label1.get_window_extent(renderer)
   1086              for tick in ticks if tick.label1.get_visible()],
   1087             [tick.label2.get_window_extent(renderer)
   1088              for tick in ticks if tick.label2.get_visible()])

File ~/.local/lib/python3.8/site-packages/matplotlib/axis.py:1085, in <listcomp>(.0)
   1083 def _get_tick_bboxes(self, ticks, renderer):
   1084     """Return lists of bboxes for ticks' label1's and label2's."""
-> 1085     return ([tick.label1.get_window_extent(renderer)
   1086              for tick in ticks if tick.label1.get_visible()],
   1087             [tick.label2.get_window_extent(renderer)
   1088              for tick in ticks if tick.label2.get_visible()])

File ~/.local/lib/python3.8/site-packages/matplotlib/text.py:910, in Text.get_window_extent(self, renderer, dpi)
    907     raise RuntimeError('Cannot get window extent w/o renderer')
    909 with cbook._setattr_cm(self.figure, dpi=dpi):
--> 910     bbox, info, descent = self._get_layout(self._renderer)
    911     x, y = self.get_unitless_position()
    912     x, y = self.get_transform().transform((x, y))

File ~/.local/lib/python3.8/site-packages/matplotlib/text.py:309, in Text._get_layout(self, renderer)
    306 ys = []
    308 # Full vertical extent of font, including ascenders and descenders:
--> 309 _, lp_h, lp_d = renderer.get_text_width_height_descent(
    310     "lp", self._fontproperties,
    311     ismath="TeX" if self.get_usetex() else False)
    312 min_dy = (lp_h - lp_d) * self._linespacing
    314 for i, line in enumerate(lines):

File ~/.local/lib/python3.8/site-packages/matplotlib/backends/backend_agg.py:259, in RendererAgg.get_text_width_height_descent(self, s, prop, ismath)
    257     texmanager = self.get_texmanager()
    258     fontsize = prop.get_size_in_points()
--> 259     w, h, d = texmanager.get_text_width_height_descent(
    260         s, fontsize, renderer=self)
    261     return w, h, d
    263 if ismath:

File ~/.local/lib/python3.8/site-packages/matplotlib/texmanager.py:335, in TexManager.get_text_width_height_descent(self, tex, fontsize, renderer)
    333 if tex.strip() == '':
    334     return 0, 0, 0
--> 335 dvifile = self.make_dvi(tex, fontsize)
    336 dpi_fraction = renderer.points_to_pixels(1.) if renderer else 1
    337 with dviread.Dvi(dvifile, 72 * dpi_fraction) as dvi:

File ~/.local/lib/python3.8/site-packages/matplotlib/texmanager.py:271, in TexManager.make_dvi(self, tex, fontsize)
    262     # Generate the dvi in a temporary directory to avoid race
    263     # conditions e.g. if multiple processes try to process the same tex
    264     # string at the same time.  Having tmpdir be a subdirectory of the
   (...)
    268     # the absolute path may contain characters (e.g. ~) that TeX does
    269     # not support.)
    270     with TemporaryDirectory(dir=Path(dvifile).parent) as tmpdir:
--> 271         self._run_checked_subprocess(
    272             ["latex", "-interaction=nonstopmode", "--halt-on-error",
    273              f"../{texfile.name}"], tex, cwd=tmpdir)
    274         (Path(tmpdir) / Path(dvifile).name).replace(dvifile)
    275 return dvifile

File ~/.local/lib/python3.8/site-packages/matplotlib/texmanager.py:241, in TexManager._run_checked_subprocess(self, command, tex, cwd)
    237     raise RuntimeError(
    238         'Failed to process string with tex because {} could not be '
    239         'found'.format(command[0])) from exc
    240 except subprocess.CalledProcessError as exc:
--> 241     raise RuntimeError(
    242         '{prog} was not able to process the following string:\n'
    243         '{tex!r}\n\n'
    244         'Here is the full report generated by {prog}:\n'
    245         '{exc}\n\n'.format(
    246             prog=command[0],
    247             tex=tex.encode('unicode_escape'),
    248             exc=exc.output.decode('utf-8'))) from exc
    249 _log.debug(report)
    250 return report

RuntimeError: latex was not able to process the following string:
b'lp'

Here is the full report generated by latex:
This is pdfTeX, Version 3.14159265-2.6-1.40.20 (TeX Live 2019/Debian) (preloaded format=latex)
 restricted \write18 enabled.
entering extended mode
(../aba7304bf2416c0194a8a27472595813.tex
LaTeX2e <2020-02-02> patch level 2
L3 programming layer <2020-02-14>
(/usr/share/texlive/texmf-dist/tex/latex/base/article.cls
Document Class: article 2019/12/20 v1.4l Standard LaTeX document class
(/usr/share/texlive/texmf-dist/tex/latex/base/size10.clo))
(/usr/share/texlive/texmf-dist/tex/latex/type1cm/type1cm.sty)

! LaTeX Error: File `type1ec.sty' not found.

Type X to quit or <RETURN> to proceed,
or enter new name. (Default extension: sty)

Enter file name:
! Emergency stop.
<read *>

l.6 \usepackage
               [utf8]{inputenc}^^M
No pages of output.
Transcript written on aba7304bf2416c0194a8a27472595813.log.
```
