### 2022-08-15 #4483 @rieder
Finally getting back to this.
@webbjj, a few suggestions:
- Could you include in README.md a short installation method (e.g. 'pip install clustertools')
- Could you also include a link to the documentation (online, e.g. on readthedocs, or as a pdf)
