### 2022-08-15 #4483 @rieder
When testing `clustertools` with an AMUSE cluster, I run into unit problems:
```python
!pip install amuse-framework amuse-masc

import clustertools as ctools
from amuse.units import units
from amuse.ext.masc import new_star_cluster

stars = new_star_cluster()
print(p.total_mass().in_(units.MSun))

cluster=ctools.StarCluster(ctype='amuse')
cluster.add_stars(p.x, p.y, p.z, p.vx, p.vy, p.vz, p.mass, p.key)
cluster.analyze(sortstars=True)
```
```
IncompatibleUnitsException: Cannot express none in kg, the units do not have the same bases
```
It looks like the total mass of the cluster is not using an AMUSE unit here, where all the stellar masses are using a unit. These can then not be compared. Adding the stars with `p.mass.value_in(units.MSun)` seems to be a workaround, but I think when `ctype` is `amuse` it would be better to use AMUSE units (inherited from the first star added?) throughout.
