### 2022-08-17 #4483 @rieder
I've created a script to compare Clustertools analysis to internal AMUSE analysis (and manual calculation), it seems they mostly agree but not entirely. Results vary with random realisations, but generally I find deviations in r10 (which is about 2 to 3 times what I would expect).
Here's the script: https://gist.github.com/rieder/5b8c40f2cff8c932c7899a1f78096952
