### 2022-08-22 #4483 @lwang-astro
@webbjj Thank for the link. It provides the correct suggestion to solve the issue.
I found two packages are required on linux:
cm-super and dvipng
For Ubuntu, I use `sudo apt install cm-super dvipng` can solve the plotting error.
