### 2022-08-24 ec8b5c8
debug adding single star particle
