### 2022-08-24 d217301
debug adding single star particle
