### 2022-08-24 f54f643
update installation docs
