+++ README.md
@@ -35,6 +35,8 @@ or
 
 pip install clustertools --install-option="--prefix=PATH"
 
+Note: For users looking to take advantage of the plotting features available in clustertools, it may be necessary to install the cm-super and dvipng packages if they aren't installed by default.
+
 # Requirements
 clustertools requires the following python packages:
 
