### 2022-08-24 d8eb316
change how stars are appended in add_stars
