### 2022-08-25 315b971
allow for float ID in add_stars
