### 2022-09-12 dc465fa
don't assume first new_gyrfalcon step is zero
