+++ clustertools/util/units.py
@@ -7,6 +7,18 @@ __author__ = "Jeremy J Webb"
 import numpy as np
 from ..util.constants import *
 
+try:
+    import amuse.units.units as u
+    from amuse.io import read_set_from_file
+except:
+    pass
+
+try:
+    from galpy.util import coords,conversion
+except:
+    import galpy.util.bovy_coords as coords
+    import galpy.util.bovy_conversion as conversion
+
 def _convert_length(x,units,cluster):
     """Convert x from units to cluster.units
 
@@ -29,10 +41,11 @@ def _convert_length(x,units,cluster):
     2022 - Written - Webb (UofT)
     """
 
-
     if units=='nbody':
-        if cluster.units=='pckms' or cluster.units=='pcmyr':
+        if cluster.units=='pckms' or cluster.units=='pcmyr' or cluster.units=='amuse':
             x*=cluster.rbar
+            if cluster.units=='amuse':
+                x= x | u.parsec
         elif cluster.units=='kpckms' or cluster.units=='kpcgyr' or cluster.units=='WDunits':
             x*=cluster.rbar/1000.0
         elif cluster.units=='galpy':
@@ -41,7 +54,10 @@ def _convert_length(x,units,cluster):
             x*=cluster.rbar/1000.0
             x=np.degrees(np.arctan2(x,cluster.zgc))
 
-    elif units=='pckms' or units=='pcmyr':
+    elif units=='pckms' or units=='pcmyr' or units=='amuse':
+        if units=='amuse':
+            x=x.value_in(u.pc)
+
         if cluster.units=='nbody':
             x/=cluster.rbar
         elif cluster.units=='kpckms' or cluster.units=='kpcgyr' or cluster.units=='WDunits':
@@ -51,20 +67,26 @@ def _convert_length(x,units,cluster):
         elif cluster.units=='radec':
             x/=1000.0
             x=np.degrees(np.arctan2(x,cluster.zgc))
+        elif cluster.units=='amuse':
+            x=x | u.parsec
 
     elif units=='kpckms' or units=='kpcgyr' or units=='WDunits':
         if cluster.units=='nbody':
             x*=1000.0/cluster.rbar
-        elif cluster.units=='pckms' or cluster.units=='pcmyr':
+        elif cluster.units=='pckms' or cluster.units=='pcmyr' or cluster.units=='amuse':
             x*=1000.0
+            if cluster.units=='amuse':
+                x= x | u.parsec
         elif cluster.units=='galpy':
             x/=cluster._ro
         elif cluster.units=='radec':
             x=np.degrees(np.arctan2(x,cluster.zgc))
 
     elif units=='galpy':
-        if cluster.units=='pckms' or cluster.units=='pcmyr':
+        if cluster.units=='pckms' or cluster.units=='pcmyr' or cluster.units=='amuse':
             x*=(1000.0*cluster._ro)
+            if cluster.units=='amuse':
+                x= x | u.parsec
         elif cluster.units=='kpckms' or cluster.units=='kpcgyr' or cluster.units=='WDunits':
             x*=cluster._ro
         elif cluster.units=='nbody':
@@ -80,9 +102,230 @@ def _convert_length(x,units,cluster):
 
             if cluster.units=='nbody':
                 x*=1000.0/cluster.rbar
-            elif cluster.units=='pckms' or cluster.units=='pcmyr':
+            elif cluster.units=='pckms' or cluster.units=='pcmyr' or cluster.units=='amuse':
                 x*=1000.0
+                if cluster.units=='amuse':
+                    x= x | u.parsec
             elif cluster.units=='galpy':
                 x/=cluster._ro
 
     return x
+
[…]
