### 2022-09-12 eed72c8
debug npop with hdf5=True
