### 2022-09-12 13f2fd4
debug npop with hdf5=True
