### 2022-09-13 6cbcd3d
debug new_gyrfalcon
