+++ tests/test_amuse_orbits.py
@@ -1,0 +1,504 @@
+import clustertools as ctools
+import numpy as np
+from galpy.orbit import Orbit
+from galpy.potential import MWPotential2014,PlummerPotential,IsochronePotential,KeplerPotential
+try:
+	from galpy.util import conversion
+except:
+	import galpy.util.bovy_conversion as conversion
+
+try:
+	import amuse.units.units as u
+except:
+	pass
+
+solar_motion=[-11.1,12.24,7.25] #Schönrich, R., Binney, J., Dehnen, W., 2010, MNRAS, 403, 1829
+solar_ro=8.275 #Gravity Collaboration, Abuter, R., Amorim, A., et al. 2020 ,A&A, 647, A59
+solar_vo=solar_ro*30.39-solar_motion[1]
+
+def test_initialize_orbit(tol=0.001):
+	cluster=ctools.load_cluster(ctype='limepy',gcname='NGC6101')
+	cluster.to_galaxy()
+	cluster.to_amuse()
+
+	o=Orbit.from_name('NGC6101',ro=solar_ro,vo=solar_vo,solarmotion=solar_motion)
+
+	ocluster=cluster.initialize_orbit()
+
+	assert np.fabs(o.x()-ocluster.x()) <= tol
+	assert np.fabs(o.y()-ocluster.y()) <= tol
+	assert np.fabs(o.z()-ocluster.z()) <= tol
+	assert np.fabs(o.vx()-ocluster.vx()) <= tol
+	assert np.fabs(o.vy()-ocluster.vy()) <= tol
+	assert np.fabs(o.vz()-ocluster.vz()) <= tol
+
+	ocluster=cluster.initialize_orbit(from_centre=True)
+
+	assert np.fabs(o.x()-ocluster.x()+cluster.xc.value_in(u.kpc)) <= tol
+	assert np.fabs(o.y()-ocluster.y()+cluster.yc.value_in(u.kpc)) <= tol
+	assert np.fabs(o.z()-ocluster.z()+cluster.zc.value_in(u.kpc)) <= tol
+	assert np.fabs(o.vx()-ocluster.vx()+cluster.vxc.value_in(u.kms)) <= tol
+	assert np.fabs(o.vy()-ocluster.vy()+cluster.vyc.value_in(u.kms)) <= tol
+	assert np.fabs(o.vz()-ocluster.vz()+cluster.vzc.value_in(u.kms)) <= tol
+
+def test_initialize_orbits(tol=0.0001):
+	
+	x=np.random.rand(1000)
+	y=np.random.rand(1000)
+	z=np.random.rand(1000)
+	vx=np.random.rand(1000)
+	vy=np.random.rand(1000)
+	vz=np.random.rand(1000)
+
+	cluster=ctools.StarCluster(units='pckms',origin='cluster')
+	cluster.add_stars(x,y,z,vx,vy,vz,m=1.,sortstars=True,analyze=True)
+
+	cluster.add_orbit(8000.,0.,0.,0.,220.,0.)
+	cluster.find_centre()
+	cluster.to_galaxy()
+	cluster.to_amuse()
+
+
+	ocluster=cluster.initialize_orbits()
+
+	np.testing.assert_allclose(8.0+x/1000.,ocluster.x(),rtol=tol)
+	np.testing.assert_allclose(y/1000.,ocluster.y(),rtol=tol)
[…]
