+++ clustertools/analysis/orbits.py
@@ -30,9 +30,15 @@ import matplotlib.pyplot as plt
 from ..util.recipes import interpolate, binmaker
 from ..util.plots import starplot,skyplot,_plot,_lplot,_scatter
 from ..util.constants import *
+from ..util.units import _convert_length,_convert_time,_convert_velocity
 
 import astropy.coordinates as coord
 import astropy.units as u
+
+try:
+    import amuse.units.units as u
+except:
+    pass
 
 def initialize_orbit(cluster, from_centre=False, ro=None, vo=None, zo = None, solarmotion=None):
     """ Initialize a galpy orbit instance for the cluster
@@ -236,11 +242,13 @@ def _integrate_orbit(
     if tfinal is None:
         tfinal=12./conversion.time_in_Gyr(ro=ro, vo=vo)
     elif cluster.units=='pckms' or cluster.units=='pcmyr':
-        tfinal/=1000.
+        tfinal/=(1000.*conversion.time_in_Gyr(ro=ro, vo=vo))
+    elif cluster.units=='amuse':
+        tfinal=tfinal.value_in(u.Gyr)/conversion.time_in_Gyr(ro=ro, vo=vo)
     elif cluster.units=='kpckms' or cluster.units=='radec' or cluster.units=='kpcgyr' or cluster.units=='WDunits':
         tfinal/=conversion.time_in_Gyr(ro=ro, vo=vo)
     elif cluster.units=='nbody':
-        tfinal*=(cluster.tbar/1000.)
+        tfinal*=((cluster.tbar/1000.)/conversion.time_in_Gyr(ro=ro, vo=vo))
 
     ts = np.linspace(0, tfinal, nt)
     o.integrate(ts, pot)
@@ -298,6 +306,8 @@ def _integrate_orbits(
         tfinal=12./conversion.time_in_Gyr(ro=ro, vo=vo)
     elif cluster.units=='pckms' or cluster.units=='pcmyr':
         tfinal/=(1000.*conversion.time_in_Gyr(ro=ro, vo=vo))
+    elif cluster.units=='amuse':
+        tfinal=tfinal.value_in(u.Gyr)/conversion.time_in_Gyr(ro=ro, vo=vo)
     elif cluster.units=='kpckms' or cluster.units=='radec' or cluster.units=='kpcgyr' or cluster.units=='WDunits':
         tfinal/=conversion.time_in_Gyr(ro=ro, vo=vo)
     elif cluster.units=='nbody':
@@ -372,6 +382,10 @@ def interpolate_orbit(
     cluster.save_cluster()
     units0,origin0, rorder0, rorder_origin0 = cluster.units0,cluster.origin0, cluster.rorder0, cluster.rorder_origin0
 
+    if cluster.units=='amuse':
+        cluster.to_pckms()
+        tfinal=tfinal.value_in(u.Myr)
+
     ts,o=_integrate_orbit(cluster, pot=pot, tfinal=tfinal, nt=nt, from_centre=from_centre, ro=ro, vo=vo,zo=zo,solarmotion=solarmotion, plot=False)
 
     if cluster.units=='radec':
@@ -379,42 +393,21 @@ def interpolate_orbit(
             cluster.to_sky()
         xgc,ygc,zgc=o.ra(ts[-1]),o.dec(ts[-1]),o.dist(ts[-1])
         vxgc,vygc,vzgc=o.pmra(ts[-1]),o.pmdec(ts[-1]),o.vlos(ts[-1])  
+
+        cluster.return_cluster(units0,origin0, rorder0, rorder_origin0)
+
     else:
[…]
