### 2022-09-16 #4483 @webbjj
@rieder - after some serious rewrites I think clustertools should now fully support AMUSE units. You would have to specify `units='AMUSE'` though when you initiate the `StarCluster`. The main issue was that I build all the arrays using numpy.append, which seemed to take an AMUSE VectorQuantity and turn it into an array of ScalarQuantities which many functions wouldn't work on. So I have just changed how the arrays are built when the input is VectorQuantity or ScalarQuantity. I have recreated all the tests I had previously written but for clusters in AMUSE units, and everything seems to check out. I also had to write some functions to give calculated variables AMUSE units if they are calculated in different units (e.g. galpy units).

@rieder and @lwang-astro  - I think I am now all caught up with your comments so far. Please let me know if I have missed something and you are waiting on me. Looking forward to more feedback if you have any. This has helped the code tremendously. Thanks so much.
