### 2022-09-20 027df90
add a return argument features to closest_star
