### 2022-10-14 #4483 @lwang-astro
@webbjj Sorry for the late of reply. Recently I am busy with teaching. Now I am continuing to test the example notebook.
I use the newest pip version of clustertools.

I got an error with
```
cluster.rcore(plot=True)
print(cluster.rc)```

The error message:
```
UnboundLocalError                         Traceback (most recent call last)
Input In [40], in <module>
----> 1 cluster.rcore(plot=True)
      2 print(cluster.rc)

File ~/.local/lib/python3.8/site-packages/clustertools/cluster/cluster.py:3151, in rcore(self, method, nneighbour, mfrac, projected, plot, **kwargs)
   3118 def ckin(
   3119     self,
   3120     mmin=None,
   (...)
   3132     projected=False,
   3133 ):
   3134     """
   3135     NAME: Find the kinematic concentration parameter ck
   3136
   3137     - see Bianchini et al. 2018, MNRAS, 475, 96
   3138
   3139     Parameters
   3140     ----------
   3141     cluster : class
   3142         StarCluster instance
   3143     mmin/mmax : float
   3144         specific mass range
   3145     nmass :
   3146         number of mass bins used to calculate alpha
   3147     rmin/rmax :
   3148         specific radial range
   3149     vmin/vmax : float
   3150         specific velocity range
-> 3151     emin/emax : float
   3152         specific energy range
   3153     kwmin/kwmax : int
   3154         specific stellar evolution type range
   3155     npop : int
   3156         population number
   3157     indx : bool
   3158         specific subset of stars
   3159     projected : bool
   3160         use projected values (default: False)
   3161
   3162     Returns
   3163     -------
   3164     ck : float
   3165         kinematic concentration
   3166
   3167     Other Parameters
   3168     ----------------
   3169     kwargs : str
   3170         key words for plotting
   3171
   3172     History
   3173     -------
   3174     2020
   3175     """
   3176     c_k = ckin(
   3177         self,
   3178         mmin=mmin,
   (...)
   3190         projected=projected,
   3191     )
   3193     self.ck=c_k

File ~/.local/lib/python3.8/site-packages/clustertools/analysis/functions.py:2146, in rcore(cluster, method, nneighbour, mfrac, projected, plot, **kwargs)
   2122 def rcore(
   2123     cluster,
   2124     method='casertano',
   (...)
   2129     **kwargs
   2130 ):
   2131     """Calculate core radius of the cluster
   2132     -- The default method (method='casertano') follows Casertano, S., Hut, P. 1985, ApJ, 298, 80 to find the core
   2133     -- An alternative metrhod (method=='isothermal') assumes the cluster is an isothermal sphere the core radius is where density drops to 1/3 central value
   2134     --- For projected core radius, the core radius is where the surface density profile drops to 1/2 the central value
   2135     --- Note that the inner mass fraction of stars used to calculate central density is set by mfrac (default 0.1 = 10%)
   2136
   2137     Parameters
   2138     ----------
   2139     cluster : class
   2140         StarCluster instance
   2141     method : string
   2142         method of calculating the core radius of a star cluster (default 'casertano')
   2143     if method =='casertano':
   2144         nneighbour : int
   2145             number of neighbours for calculation local densities
-> 2146     if method=='isothermal':
   2147         mfrac : float
   2148             inner mass fraction to be used to establish the central density
   2149     projected : bool
   2150         use projected values (default: False)
   2151     plot : bool
   2152         plot the density profile and mark the core radius of the cluster (default: False)
   2153     Returns
   2154     -------
   2155     rc : float
   2156         core radius
   2157
   2158     Other Parameters
   2159     ----------------
   2160     None
   2161
   2162     History
   2163     -------
   2164     2021 - Written - Webb (UofT)
   2165     2022 - Written - Webb (UofT) - add method='casertano'
   2166     """
   2167     cluster.save_cluster()
   2168     units0,origin0, rorder0, rorder_origin0 = cluster.units0,cluster.origin0, cluster.rorder0, cluster.rorder_origin0

UnboundLocalError: local variable 'nrad' referenced before assignment
```
