### 2022-10-14 5d8e941
debug rcore plotting
