### 2022-10-14 d6525ca
debug rcore plotting
