### 2022-10-14 #4483 @webbjj
My apologies @lwang-astro , it looks like I broke the plotting feature when adding your suggested method of calculating rcore. Should be fixed now in the newest pip (clustertools-1.0.4) or git hub.
