### 2022-11-18 826f6f4
set default plotting properties
