### 2022-11-29 #4483 @lwang-astro
@webbjj @dfm Sorry for the late of reply. Now I am continuing to check the tool. For the previous problem of plotting bug.
`Latex Error: File 'type1cm.sty' not found. `

I test the code on the other machine and found that only installing cm-super and dvipng are not enough, for ubuntu: `sudo apt install cm-super dvipng`. The texlive and texlive-extra also needed to be installted together. I am not sure which function in the clustertool requires texlive and texlive-extra. Generally the default matplotlib should support XML mathmatical symbols without these additional packages.

I have tested the samples in Example Notebooks except the last one "[Orbits and Tidal Tails](https://clustertools.readthedocs.io/en/latest/notebooks/orbits_and_tails.html)" , since I have no available sample data to test it directly. For other example sources, they work if the plotting problem is solved.
