+++ docs/source/conf.py
@@ -25,7 +25,7 @@ else:
 author = 'Jeremy J. Webb'
 
 # The full version, including alpha/beta/rc tags
-release = '1.0.5'
+release = '1.0.6'
 
 
 # -- General configuration ---------------------------------------------------
