### 2022-12-02 92cbc4d
debug latex error when plotting
