### 2022-12-02 #4483 @webbjj
Thanks @lwang-astro !

- Just to confirm, installing texlive and texlive-extra fixes the plotting issue? If so I will add notes to the documentation, however I will endeavour to try and figure this out. Its hard because I can't reproduce it and its not clear to me why matplotlib isn't working within clustertools.
- For the orbits and tails notebook, the data you would need (pal5.dat) can be found in clustertools/docs/source/notebooks if you cloned the git repo. Otherwise I could try and send it directly.
