### 2022-12-05 #4483 @lwang-astro
@webbjj For ubuntu, I think that installation of texlive is necessay to show latex symbol if matplotlib require latex compiler. For some reason which I don't know in detail, some ploting function in clustertools require type1cm.sty, which is only available in texlive-extra. Thus, to make the plotting function work correctly, both are needed to be installed.
