### 2022-12-08 #4483 @webbjj
Thanks @lwang-astro - ok I have added a note to the installation instructions. However I also "maybe" have fixed the issue. I was relying on seaborn for a colour palette, and seaborn changes some of the matplotlib parameters. There was also a galpy plotting function that was changing matplotlib parameters too. So I have removed these dependencies and try to import matplotlib last everywhere. Its hard to tell, because I can't reproduce the error. But as far as I can tell the default matplotlib parameters are now the same before and after importing clustertools.
