### 2022-12-15 #4483 @rieder
Sorry for taking a while to get back to this...
I'm looking at the tests now, and these seem to fail, with varying errors:
- first, they need some testing package installed (I'm using pytest).
- they then complain about astropy missing
- also, they need amuse (or at least amuse-framework)
- then some need limepy, but the version on PyPI doesn't seem to work (`from limepy import limepy` fails)

I am not sure if these are all the errors in the tests.
