### 2022-12-15 #4483 @webbjj
Hi @rieder , thanks for the feedback.

- I use pytest as well, so I am not sure what testing package error you are getting. Can you elaborate?
- I don't list astropy, amuse-framework, and limepy as essential packages because someone can use clustertools without having them installed. However in order to test things like loading an amuse particle set, a limepy model, or an astropy table I need to be able to import those. I also use limepy to setup mock clusters in the tests. I wasn't anticipating someone wanting to run the tests themselves, but I understand now that it would be a way to test that the code is working. Is there a better practice for setting up the tests for these kinds of things?
