### 2022-12-15 #4483 @rieder
```
ImportError while importing test module '/Users/rieder/Review/clustertools/tests/test_loaders.py'.
Traceback:
/opt/local/Library/Frameworks/Python.framework/Versions/3.10/lib/python3.10/importlib/__init__.py:126: in import_module
tests/test_loaders.py:13: in <module>
    from limepy import limepy
E   ImportError: cannot import name 'limepy' from 'limepy' (/Users/rieder/Review/clustertools/env/lib/python3.10/site-packages/limepy/__init__.py)
```
this is an example of an error I get from limepy not importing correctly in one of the tests. It prevents the tests from running at all.
