### 2022-12-22 3da9e3b
remove mandatory dependence of tests on limepy, amuse, and astropy
