+++ tests/test_amuse_functions.py
@@ -2,14 +2,13 @@ import clustertools as ctools
 from clustertools.analysis.functions import tpl_func
 from scipy.optimize import curve_fit
 
+import pytest
+
 import numpy as np
 from galpy.orbit import Orbit
 from galpy.potential import NFWPotential,MWPotential2014,rtide,evaluateDensities
 from galpy.df import isotropicNFWdf
 from galpy.util import bovy_conversion
-
-import limepy
-from limepy import limepy
 
 try:
     from galpy.util import coords,conversion
@@ -19,8 +18,9 @@ except:
 
 try:
 	import amuse.units.units as u
+	noamuse=False
 except:
-	pass
+	noamuse=True
 
 solar_motion=[-11.1,12.24,7.25] #Schönrich, R., Binney, J., Dehnen, W., 2010, MNRAS, 403, 1829
 solar_ro=8.275 #Gravity Collaboration, Abuter, R., Amorim, A., et al. 2020 ,A&A, 647, A59
@@ -28,8 +28,10 @@ solar_vo=solar_ro*30.39-solar_motion[1]
 solar_zo=0.0208 #Bennett & Bovy 2019
 to=conversion.time_in_Gyr(ro=solar_ro,vo=solar_vo)
 
+@pytest.mark.skipif(noamuse, reason='amuse required to run this test')
 def test_find_centre_of_density(tol=0.01):
-	cluster=ctools.load_cluster(ctype='limepy',gcname='NGC6101')
+	cluster=ctools.load_cluster(ctype='snapshot',filename='ngc6101_pckms_cluster.dat',units='pckms',origin='cluster')
+	cluster.add_orbit(-2005.2100994789871, -9348.1814843660959, -3945.4681762489472, -296.18121334354328, 82.774301940161507, -190.84753679996979)
 	cluster.to_galaxy()
 	cluster.to_kpckms()
 	o=Orbit.from_name('NGC6101',ro=solar_ro,vo=solar_vo,solarmotion=solar_motion)
@@ -47,8 +49,10 @@ def test_find_centre_of_density(tol=0.01):
 	assert np.fabs(1.0-vyc/(o.vy() | u.kms)) <= tol
 	assert np.fabs(1.0-vzc/(o.vz() | u.kms)) <= tol
 
+@pytest.mark.skipif(noamuse, reason='amuse required to run this test')
 def test_find_centre_of_density_casertano(tol=0.01):
-	cluster=ctools.load_cluster(ctype='limepy',gcname='NGC6101')
+	cluster=ctools.load_cluster(ctype='snapshot',filename='ngc6101_pckms_cluster.dat',units='pckms',origin='cluster')
+	cluster.add_orbit(-2005.2100994789871, -9348.1814843660959, -3945.4681762489472, -296.18121334354328, 82.774301940161507, -190.84753679996979)
 	cluster.to_galaxy()
 	cluster.to_kpckms()
 	o=Orbit.from_name('NGC6101',ro=solar_ro,vo=solar_vo,solarmotion=solar_motion)
@@ -66,9 +70,10 @@ def test_find_centre_of_density_casertano(tol=0.01):
 	assert np.fabs(1.0-vyc/(o.vy() | u.kms)) <= tol
 	assert np.fabs(1.0-vzc/(o.vz() | u.kms)) <= tol
 
-
+@pytest.mark.skipif(noamuse, reason='amuse required to run this test')
 def test_find_centre_of_mass(tol=0.01):
[…]
