### 2022-12-22 15c71cf
remove import astropy when not needed
