### 2022-12-22 35228c5
remove amuse dependence in tests
