+++ tests/test_operations.py
@@ -25,7 +25,11 @@ def scale_test(init,final,tscale,mscale,xscale,vscale):
 	#assume arrays are t,mass,x,y,z,vx,vy,vz,xc,yc,zc,vxc,vyc,vzc,xgc,ygc,zgc,vxgc,vygc,vzgc
 
 	for i in range(0,len(init)):
-		print(i,init[i],final[i])
+		if i==0 or i>=8:
+			print(i,init[i],final[i],final[i]/init[i],tscale,mscale,xscale,vscale)
+		else:
+			print(i,init[i][0],final[i][0],final[i][0]/init[i][0],tscale,mscale,xscale,vscale)
+
 		if i==0:
 			assert np.all(np.fabs(final[i]/init[i]-tscale) < 1e-10)
 		elif i==1:
@@ -33,8 +37,8 @@ def scale_test(init,final,tscale,mscale,xscale,vscale):
 		elif (i>=2 and i<=4) or (i>=8 and i<=10) or (i>=14 and i<=16):
 			assert np.all(np.fabs(final[i]/init[i]-xscale) < 1e-10)
 		elif (i>=5 and i<=7) or (i>=11 and i<=13) or (i>=17 and i<=19):
-			print(i,final[i],init[i],vscale)
-			assert np.all(np.fabs(final[i]/init[i]-vscale) < 1e-10)
+			print(i,vscale,np.all(np.fabs(final[i]/init[i]-vscale) < 1e-9))
+			assert np.all(np.fabs(final[i]/init[i]-vscale) < 1e-9)
 
 def test_to_pckms(tol=0.0001,ro=solar_ro,vo=solar_vo):
 
@@ -70,6 +74,8 @@ def test_to_pckms(tol=0.0001,ro=solar_ro,vo=solar_vo):
 
 def test_to_pcmyr(tol=0.0001,ro=solar_ro,vo=solar_vo):
 
+	np.random.seed(5)
+
 	tbar,rbar,vbar,zmbar=2.1,5.3,2.5,7553.7
 
 	init_units=['kpckms','galpy','nbody','WDunits','pckms','kpcgyr']
@@ -101,6 +107,7 @@ def test_to_pcmyr(tol=0.0001,ro=solar_ro,vo=solar_vo):
 		final=[cluster.tphys,cluster.m,cluster.x,cluster.y,cluster.z,cluster.vx,cluster.vy,cluster.vz,cluster.xc,cluster.yc,cluster.zc,cluster.vxc,cluster.vyc,cluster.vzc,cluster.xgc,cluster.ygc,cluster.zgc,cluster.vxgc,cluster.vygc,cluster.vzgc]
 		scale_test(init,final,tscales[i],mscales[i],xscales[i],vscales[i])
 
+	assert 1==0
 
 
 def test_to_kpckms(tol=0.0001,ro=solar_ro,vo=solar_vo):
@@ -790,82 +797,84 @@ def test_to_units(tol=0.01,ro=solar_ro,vo=solar_vo):
 	assert np.all(np.fabs(os.vlos()/cluster.vlos-1.) < 1e-10)
 	assert np.all(np.fabs(os.vlos()/cluster.vz-1.) < 1e-10)
 
-	tbar,rbar,vbar,zmbar=2.1,5.3,2.5,7553.7
+	if not noamuse:
 
-	init_units=['pckms','kpckms','galpy','nbody','WDunits','pcmyr','kpcgyr']
-	tscales=[1.,1000,(1000.*conversion.time_in_Gyr(ro=ro,vo=vo)),tbar,1000.0,1.,1000.]
-	xscales=[1.,1000,ro*1000,rbar,1000.,1.,1000.]
[…]
