+++ tests/test_loaders.py
@@ -2,6 +2,7 @@ import clustertools as ctools
 import numpy as np
 
 import pytest
+import os
 
 try:
 	from astropy.table import QTable
@@ -28,6 +29,8 @@ try:
 	nolimepy=False
 except:
 	nolimepy=True
+
+nosim=np.invert(os.path.isfile('../docs/source/notebooks/nbody6_sim/OUT3'))
 
 solar_motion=[-11.1,12.24,7.25] #Schönrich, R., Binney, J., Dehnen, W., 2010, MNRAS, 403, 1829
 solar_ro=8.275 #Gravity Collaboration, Abuter, R., Amorim, A., et al. 2020 ,A&A, 647, A59
@@ -449,6 +452,7 @@ def check_params(cluster,ctype,units,origin,projected,**kwargs):
 
 	return True
 
+@pytest.mark.skipif(nosim, reason='A Nbody6 Simulation must be saved at ../docs/source/notebooks/nbody6_sim/')
 def test_nbody6():
 	wdir='../docs/source/notebooks/nbody6_sim/'
 
