### 2022-12-23 6caf0f4
remove astropy dependence from tests
