+++ tests/test_amuse_orbits.py
@@ -26,7 +26,7 @@ def test_initialize_orbit(tol=0.001):
 	cluster.to_galaxy()
 	cluster.to_amuse()
 
-	o=Orbit.from_name('NGC6101',ro=solar_ro,vo=solar_vo,solarmotion=solar_motion)
+	o=Orbit([246.44999999999987, -72.201999999999998, 14.450000000000003, 1.7557999999999991, -0.25770000000000326, 366.32999999999998],radec=True,ro=solar_ro,vo=solar_vo,solarmotion=solar_motion)
 
 	ocluster=cluster.initialize_orbit()
 
@@ -109,7 +109,7 @@ def test_interpolate_orbit(tol=0.1,ro=solar_ro,vo=solar_vo):
 
 	x,y,z,vx,vy,vz=cluster.interpolate_orbit(pot=MWPotential2014,tfinal=1. | u.Gyr, nt=1000)
 
-	o=Orbit.from_name('NGC6101',ro=solar_ro,vo=solar_vo,solarmotion=solar_motion)
+	o=Orbit([246.44999999999987, -72.201999999999998, 14.450000000000003, 1.7557999999999991, -0.25770000000000326, 366.32999999999998],radec=True,ro=solar_ro,vo=solar_vo,solarmotion=solar_motion)
 	ts=np.linspace(0,1./conversion.time_in_Gyr(ro=ro,vo=vo))
 	o.integrate(ts,MWPotential2014)
 
@@ -260,7 +260,7 @@ def test_orbital_path(tol=0.1,ro=solar_ro,vo=solar_vo):
 	    plot=False,
 	)
 
-	o=Orbit.from_name('NGC6101',ro=solar_ro,vo=solar_vo,solarmotion=solar_motion)
+	o=Orbit([246.44999999999987, -72.201999999999998, 14.450000000000003, 1.7557999999999991, -0.25770000000000326, 366.32999999999998],radec=True,ro=solar_ro,vo=solar_vo,solarmotion=solar_motion)
 	ts = np.linspace(0, (-1.0* tfinal.value_in(u.Gyr))/ conversion.time_in_Gyr(ro=ro, vo=vo), 1000)
 	o.integrate(ts,MWPotential2014)
 
@@ -285,7 +285,7 @@ def test_orbital_path(tol=0.1,ro=solar_ro,vo=solar_vo):
 	assert np.fabs(cluster.orbit.vy(ts[-1])-o.vy(ts[-1])) <= tol
 	assert np.fabs(cluster.orbit.vz(ts[-1])-o.vz(ts[-1])) <= tol
 
-	o=Orbit.from_name('NGC6101',ro=solar_ro,vo=solar_vo,solarmotion=solar_motion)
+	o=Orbit([246.44999999999987, -72.201999999999998, 14.450000000000003, 1.7557999999999991, -0.25770000000000326, 366.32999999999998],radec=True,ro=solar_ro,vo=solar_vo,solarmotion=solar_motion)
 	ts = np.linspace(0, (1.0* tfinal.value_in(u.Gyr))/ conversion.time_in_Gyr(ro=ro, vo=vo), 1000)
 	o.integrate(ts,MWPotential2014)
 
@@ -355,7 +355,7 @@ def test_orbital_path_match(tol=0.1,ro=solar_ro,vo=solar_vo):
 	tfinal=0.1
 	nt=1000
 
-	o=Orbit.from_name('NGC6101',ro=solar_ro,vo=solar_vo,solarmotion=solar_motion)
[…]
