+++ clustertools/__init__.py
@@ -1,4 +1,4 @@
-__version__= "1.0.7"
+__version__= "1.0.8"
 from .analysis import *
 from .cluster import *
 from .io import *
