### 2022-12-23 #4483 @webbjj
Thanks @rieder , I was unaware of the skip option in pytest. I have now rewritten the tests so there is no dependence on astropy, limepy, or amuse unless absolutely necessary. In those cases the tests are skipped if those packages are not installed.
