### 2023-01-29 #4483 @lwang-astro
Sorry for the long delay. Now I am testing the functions in orbits_and_tails.ipynb using the version 1.0.8.
Here are the two problems I have found:

1. The absolute path does not work for the keyword argument `snapdir` of the `load_cluster` function:
```
cluster = cts.load_cluster('snapshot',snapdir='/home/lwang/src/python/cluster/', filename='pal5.dat', units='kpckms',origin='galaxy')
```
This return `NO FILE FOUND: ./, /home/lwang/src/python/cluster/, pal5.dat`

For a relative path, it is fine:
```
cluster = cts.load_cluster('snapshot',snapdir='cluster/', filename='pal5.dat', units='kpckms',origin='galaxy')
```

2.  When I try
```
cluster = cts.load_cluster('limepy',gcname='Pal5',origin='galaxy',units='kpckms')
cts.starplot(cluster,do_centre=True)
```
It returns an error of load limepy (I have installed limepy with pip):
```

NameError                                 Traceback (most recent call last)
Input In [29], in <module>
----> 1 cluster = cts.load_cluster('limepy',gcname='Pal5',origin='galaxy',units='kpckms')
      2 cts.starplot(cluster,do_centre=True)

File ~/.local/lib/python3.8/site-packages/clustertools/io/load.py:275, in load_cluster(ctype, units, origin, ofile, orbit, filename, particles, load_function, **kwargs)
    272     cluster = _get_galpy_orbits(particles, units=units, origin=origin, ofile=ofile, **kwargs)
    274 elif ctype == 'limepy':
--> 275     cluster = _get_limepy(units=units, origin=origin, orbit=orbit, ofile=ofile, **kwargs)
    277 else:
    278     print("NO CTYPE GIVEN")

File ~/.local/lib/python3.8/site-packages/clustertools/io/limepydf.py:92, in _get_limepy(units, origin, orbit, ofile, advance, **kwargs)
     90 source = kwargs.pop("source", "default")
     91 mbar = kwargs.pop("mbar", 1.)
---> 92 cluster = _get_cluster(gcname, source, mbar, **kwargs)
     94 if cluster.units!=units:
     95     if units=='nbody': cluster.reset_nbody_scale(rvirial=True)

File ~/.local/lib/python3.8/site-packages/clustertools/io/limepydf.py:764, in _get_cluster(gcname, source, mbar, params, **kwargs)
    760     cluster = _get_deBoer_cluster(ddata, gcname, mbar, **kwargs)
    761 elif (source == "default" or "harris" in source or "Harris" in source) and (
    762     gcname in hname or gcname in hname2
    763 ):
--> 764     cluster = _get_harris_cluster(hdata, gcname, mbar, **kwargs)
    765 else:
    766     print('No match: ',source,gcname, gcname in dname, gcname in hname, gcname in hname2)

File ~/.local/lib/python3.8/site-packages/clustertools/io/limepydf.py:885, in _get_harris_cluster(data, gcname, mbar, **kwargs)
    882 w0 = c_to_w0(c)
    883 N = int(kwargs.get("N", mgc / mbar))
--> 885 cluster = _sample_limepy(g=1.0, phi0=w0, M=mgc, rt=rl, N=N)
    887 cluster.units='pckms'
    888 cluster.bunits='pckms'

File ~/.local/lib/python3.8/site-packages/clustertools/io/limepydf.py:191, in _sample_limepy(g, model, **kwargs)
    189 if "rt" in kwargs:
    190     rt = float(kwargs.get("rt"))
--> 191     lmodel = limepy(phi0, g, M=M, rt=rt, project=project)
    192 elif "rv" in kwargs:
    193     rv = float(kwargs.get("rv"))

NameError: name 'limepy' is not defined

```
