### 2023-01-30 e49616a
fix confusion between wdir and snapdir
