+++ clustertools/io/nbody6pp.py
@@ -57,7 +57,14 @@ def _get_nbody6pp(conf3, bev82=None, sev83=None, snap40=None, ofile=None, advanc
 
     initialize = kwargs.get("initialize", False)
     nsnap = kwargs.pop("nsnap", 0)
-    wdir = kwargs.get("wdir", './')
+    wdir = kwargs.get("wdir", "")
+    if wdir is not "":
+        if wdir[-1] != '/':
+            wdir+='/'
+    snapdir = kwargs.get("snapdir", "")
+    if snapdir is not "":
+        if snapdir[-1] != '/':
+            snapdir+='/'
     deltat=kwargs.get('deltat',1)
     dtout=kwargs.get('dtout',deltat)
 
