+++ clustertools/__init__.py
@@ -1,4 +1,4 @@
-__version__= "1.0.8"
+__version__= "1.0.9"
 from .analysis import *
 from .cluster import *
 from .io import *
