### 2023-01-30 54ad48c
debug snapdir with nbody6
