### 2023-01-31 #4483 @lwang-astro
> Thanks for all the testing @lwang-astro. Version 1.0.9 now uploaded to pip and pushed to github.
>
> First, good catch on playing with the path being given to snapdir. In hindsight snapdir is poorly explained in the function's description as its meant to be an add on to the working directory wdir. So for example if the bulk of your simulation results are in './sim' but then the snapshots are in a subdirectory 'snaps', then wdir='./sim' and snapdir='snaps' (assuming you script is a directory above sim). However the brief description in load_function makes snapdir and wdir seem like the same thing. I have altered the text description in load_function, but perhaps most importantly I have made it so a user could set wdir and/or snapdir and it should now look in the right place. In the tests directory, within test_loaders.py there is now a test called test_filepaths that cycles through all the combinations of wdir and snapdir I could think of. It won't run on any machine but mine since I test the root path, but it hopefully tests other variations. Either way, I believe your first line should run now.
>
> Second, for the limpy issue, I can't imagine why this is happening if you have limepy installed. It seems like the following code is failing:
>
> ```
> try:
>     from limepy import limepy,spes,sample
> except:
>     pass
> ```
>
> In the tests directory, what happens when you run
>
> `pytest -v test_loaders.py -k test_limepy`
>
> Does it run and pass or does it skip because it can't find limepy? If the latter then `from limepy import limepy,spes,sample` is failing (I think). Do you import limepy a different way?
>
> Thanks again for your help.

This test did not work. Here is the error message:

```
================================================= test session starts ==================================================
platform linux -- Python 3.8.10, pytest-6.2.5, py-1.11.0, pluggy-1.0.0 -- /usr/bin/python3
cachedir: .pytest_cache
rootdir: /home/lwang/code/clustertools
collected 11 items / 10 deselected / 1 selected

test_loaders.py::test_limepy SKIPPED (limepy required to run this test)                                          [100%]

========================================== 1 skipped, 10 deselected in 14.35s ==========================================

```
When I manually do `from limepy import limepy`, it indeeds fails with the following error message.

```
In [1]: from limepy import limepy, spes, sample
ImportError                               Traceback (most recent call last)
Input In [1], in <module>
----> 1 from limepy import limepy, spes, sample

ImportError: cannot import name 'limepy' from 'limepy' (/home/lwang/.local/lib/python3.8/site-packages/limepy/__init__.py)
```

I guess one difference between my limepy version and yours is that I have used `pip3` to install limepy.
