### 2023-01-31 #4483 @webbjj
Hi @lwang-astro

Is there any chance you pip installed limepy (https://pypi.org/project/limepy/) instead of astro-limepy (https://pypi.org/project/astro-limepy/). The naming issue is quite unfortunate, but Mark Gieles' code is astro-limepy, but then behaves as if it is called limepy once pip installed.

Jeremy
