### 2023-01-31 #4483 @lwang-astro
Thanks. Indeed I have made this mistake. After pip install Astro-limepy. I got another error from limepy itself (not related to clustertools). I guess a bug exists in the current version of limepy.

In [1]: from limepy import limepy, spes, sample
Traceback (most recent call last):

  File ~/.local/lib/python3.8/site-packages/IPython/core/interactiveshell.py:3251 in run_code
    exec(code_obj, self.user_global_ns, self.user_ns)

  Input In [1] in <module>
    from limepy import limepy, spes, sample

  File ~/.local/lib/python3.8/site-packages/limepy/spes.py:849
    DF = numpy.zeros(max(len(r),len(v)))
                                        ^
TabError: inconsistent use of tabs and spaces in indentation

发件人: ***@***.***>
发送时间: 2023年1月31日 19:47
主题: Re: [openjournals/joss-reviews] [REVIEW]: clustertools: A Python Package for Analyzing Star Cluster Simulations (Issue #4483)

Hi @lwang-astro<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Flwang-astro&data=05%7C01%7C%7Cbb0dc9c4391044ec225008db0380f2fe%7C84df9e7fe9f640afb435aaaaaaaaaaaa%7C1%7C0%7C638107624586626048%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=tYReCJORAgLZLVORmHyBdTzcJteOf7S4BKLi3nA3SyA%3D&reserved=0>

Is there any chance you pip installed limepy (https://pypi.org/project/limepy/<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fpypi.org%2Fproject%2Flimepy%2F&data=05%7C01%7C%7Cbb0dc9c4391044ec225008db0380f2fe%7C84df9e7fe9f640afb435aaaaaaaaaaaa%7C1%7C0%7C638107624586626048%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=2qMVNgv%2BgEs8lhxfnFN7QwXmiLgeJDFJgvj10ZWP1Ho%3D&reserved=0>) instead of astro-limepy (https://pypi.org/project/astro-limepy/<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fpypi.org%2Fproject%2Fastro-limepy%2F&data=05%7C01%7C%7Cbb0dc9c4391044ec225008db0380f2fe%7C84df9e7fe9f640afb435aaaaaaaaaaaa%7C1%7C0%7C638107624586781636%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=1j90xd%2BcTTGU1DkKGzBSRjg9azlG%2BW7PyMDx5mAQCyo%3D&reserved=0>). The naming issue is quite unfortunate, but Mark Gieles' code is astro-limepy, but then behaves as if it is called limepy once pip installed.

Jeremy

―
