### 2023-01-31 b09f8f9
debug amuse array issue with test_amuse_orbits
