### 2023-02-20 #4483 @webbjj
Hi @lwang-astro , this is very strange as I can't replicate it. The code snippet you have above runs perfectly fine for me. You are correct that limepy is not imported in load.py, but it is in limepydf.py where I think this error is being flagged.

Do you have the latest version of clustertools? 1.011? I noticed my line number is off by one from yours. It could be nothing, but I was able to replicate the error if I just had `import limepy` in limepydf.py instead of `from limepy import limepy,spes,sample`. Somewhere along the line I think limepy changed from having to do:

`import limepy`
`from limepy import limepy,spes,sample`

to just

`from limepy import limepy,spes,sample`

So maybe with an updated version of clustertools things will work? Perhaps I removed an import limepy somewhere for this exact reason.

Another thing to check would be that limepy runs on its own. So does the following code work for you?
`>>> from limepy import limepy
>>> model=limepy(1,1)
>>> print(model.g)
1`

Jeremy
