### 2023-02-20 #4483 @lwang-astro
Hi @webbjj , thank you for the suggestion. I have tried your script and found the problem. I have installed astro-limepy via pip3 before. In this version, limepy is a module, so the error occurs. Now I removed it and use the zipped version you provided before. It seems the error disappears. However, the results seems a bit strange. See the attached snapshot:
![image](https://user-images.githubusercontent.com/8628138/220107743-2bc618e0-1352-4c2f-90ee-91c4f067c649.png)

I expect that the plot shows a tidal tail, but it gave a cluster. This is different from the original plot saved in the notebook. Is this a correct output consistent with your original design?
