### 2023-02-21 #4483 @webbjj
Thanks @lwang-astro . Ok so this is working properly now. LIMEPY just generates a lowered isothermal model for a given GC, and what. clustertools does is take the best fit model from either de Boer et al 2019 (if available) or Harris (2010) and turns it into a StarCluster class. LIMEPY won't include any tails or anything like that.

I think it was perhaps confusing on my part to use an Nbody6 Pal 5 Model (with tails) to show how Nbody6 loading works in the documentation AND Pal 5 in the limepy example as well. I should have just used another cluster. So I will update the documentation now.
