### 2023-02-21 11f6131
update notebook examples
