### 2023-04-15 #4483 @lwang-astro
@dfm, Sorry for the long delay of the processing. I have been in communication with @webbjj regarding the issues I experienced while checking the sample Python scripts. I have now tested all the functions and they are working as expected after the issues were resolved. In our latest communication, @webbjj mentioned that he would update the documentation to align it with the sample script. As this is a minor fix, I can now proceed to finish the review and approve the code publication.
