### 2023-04-16 #4483 @webbjj
Thanks for the update @lwang-astro . Apologies if you were waiting for me to confirm the change in the documentation. Now the example LIMEPY cluster is NGC 6101, and I have tried to make it a little clearer what is happening with the LIMEPY case in https://clustertools.readthedocs.io/en/latest/notebooks/setup.html. So it should hopefully not give the impression that one would expect and kind of tidal tails in the setup.

In most cases I have now changed the example cluster to be NGC 6101, as to avoid confusion with the Nbody6 Pal 5 example that has tails.
