### 2023-05-08 dfc069b
Update README.md
Use syntax highlighting for installation instructions
