### 2023-05-15 #4483 @rieder
I found it easiest to do my review (especially the "functionality" and "tests" part) using a notebook that extensively tests clustertools with AMUSE data - it's not finished but @webbjj you can already find it at https://github.com/rieder/clustertools_amuse_test.
I find some weird behaviours but I think most of this can be easily fixed or worked around.
Some of the issues so far seem to be unexpected behaviour (e.g. calling the same method a second time giving a different result) and inconsistent coordinate system use (center-of-mass seems to ignore galactic coordinates?).
