### 2023-05-15 #4483 @rieder
I find some bugs/inconsistencies/things that are unclear to me, but I don't think this should keep the review going on any longer than it already has.
I think Clustertools is a very useful, well-documented piece of code and recommend it for publication.
And my deepest apologies to @webbjj and all of you at JOSS for the loooooong delay...
