+++ paper/paper.bib
@@ -10,7 +10,7 @@
 @ARTICLE{galpy,
        author = {{Bovy}, Jo},
         title = "{galpy: A python Library for Galactic Dynamics}",
-      journal = {\apjs},
+      journal = {Astrophysical Journal, Supplement},
      keywords = {galaxies: general, galaxies: kinematics and dynamics, Galaxy: fundamental parameters, Astrophysics - Astrophysics of Galaxies, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2015,
         month = feb,
@@ -29,7 +29,7 @@ archivePrefix = {arXiv},
 @ARTICLE{limepy,
        author = {{Gieles}, Mark and {Zocchi}, Alice},
         title = "{A family of lowered isothermal models}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {methods: analytical, methods: numerical, stars: kinematics and dynamics, globular clusters: general, open clusters and associations: general, galaxies: star clusters: general, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies, Astrophysics - Solar and Stellar Astrophysics},
          year = 2015,
         month = nov,
@@ -55,7 +55,7 @@ archivePrefix = {arXiv},
 @ARTICLE{nbody6pp,
        author = {{Wang}, Long and {Spurzem}, Rainer and {Aarseth}, Sverre and {Nitadori}, Keigo and {Berczik}, Peter and {Kouwenhoven}, M.~B.~N. and {Naab}, Thorsten},
         title = "{NBODY6++GPU: ready for the gravitational million-body problem}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {methods: numerical, globular clusters: general, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2015,
         month = jul,
@@ -87,7 +87,7 @@ archivePrefix = {arXiv},
 @ARTICLE{petar,
        author = {{Wang}, Long and {Iwasawa}, Masaki and {Nitadori}, Keigo and {Makino}, Junichiro},
         title = "{PETAR: a high-performance N-body code for modelling massive collisional stellar systems}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {methods: numerical, software: simulations, globular clusters: general, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2020,
         month = sep,
@@ -105,7 +105,7 @@ archivePrefix = {arXiv},
 @ARTICLE{mcluster,
        author = {{K{\"u}pper}, Andreas H.~W. and {Maschberger}, Thomas and {Kroupa}, Pavel and {Baumgardt}, Holger},
         title = "{Mass segregation and fractal substructure in young massive clusters - I. The McLuster code and method calibration}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {methods: data analysis, Magellanic Clouds, galaxies: star clusters: individual: R136, Astrophysics - Astrophysics of Galaxies, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2011,
         month = nov,
