+++ tests/test_amuse_functions.py
@@ -18,6 +18,7 @@ except:
 
 try:
 	import amuse.units.units as u
+    from amuse.units.quantities import ScalarQuantity,VectorQuantity
 	noamuse=False
 except:
 	noamuse=True
