+++ tests/test_amuse_orbits.py
@@ -11,6 +11,8 @@ except:
 
 try:
 	import amuse.units.units as u
+    from amuse.units.quantities import ScalarQuantity,VectorQuantity
+
 	noamuse=False
 except:
 	noamuse=True
