+++ clustertools/cluster/cluster.py
@@ -26,9 +26,9 @@ from galpy.orbit import Orbit
 from ..util.constants import *
 
 try:
-    from amuse.lab import *
     import amuse.units.units as u
     from amuse.datamodel import Particles
+    from amuse.units.quantities import ScalarQuantity,VectorQuantity
     from ..util.units import _convert_amuse,_convert_length,_convert_velocity
     _hasamuse=True
 except:
@@ -127,11 +127,19 @@ class StarCluster(object):
     ):
 
         # Age of cluster
-        self.tphys = tphys
+        if (ctype == 'amuse' and _hasamuse and units is None) or (units == 'amuse' and _hasamuse) :
+            self.tphys = tphys | u.Myr
+        else:
+            self.tphys = tphys
+
         self.dt = None
 
         # Units and origin
         self.units = units
+
+        if ctype == 'amuse' and _hasamuse and units is None:
+            self.units='amuse'
+
         self.bunits = units
         self.origin = origin
         self.units_init=units
@@ -459,7 +467,7 @@ class StarCluster(object):
         #Check for AMUSE units:
         isamuse=False
 
-        if _hasamuse and (self.units=='amuse' or (self.units is None and self.ctype=='amuse')):
+        if _hasamuse:
             if isinstance(x,VectorQuantity):
                 stars=Particles(len(x))
                 stars.x,stars.y,stars.z=x,y,z
@@ -665,9 +673,9 @@ class StarCluster(object):
             self.vlos = np.append(self.vlos, vz)
 
 
-        if isamuse: 
+        if isamuse and (self.units_init is None or self.units_init == 'amuse'):
             self.to_amuse()
-            self.units_init='amuse'
+
         if analyze: self.analyze(sortstars=sortstars)
 
 
@@ -1445,7 +1453,7 @@ class StarCluster(object):
 
         2020 - Written - Webb (UofT)
         """
-        analyze(self, sortstars = sortstars, projected=projected)
+        self.analyze(sortstars = sortstars, projected=projected)
 
     def key_params(self, do_order=True, projected=True):
         """Call analyze with key_params for backwards compatibility
@@ -1468,7 +1476,7 @@ class StarCluster(object):
         2020 - Written - Webb (UofT)
         """
 
-        analyze(self,sortstars=do_order, projected=projected)
+        self.analyze(sortstars=do_order, projected=projected)
 
 
     def sortstars(self, projected=True):
@@ -2168,77 +2176,9 @@ class StarCluster(object):
             nsigma=nsigma,nsphere=nsphere,density=density,
             rmin=rmin,rmax=rmax,nmax=nmax,method=method,nneighbour=nneighbour)
 
-        if self.origin=='centre':
+        self._set_centre(xc,yc,zc,vxc,vyc,vzc,reset_centre=reset_centre)
 
-            warning=False
-            if xc!=0.0 or yc!=0.0 or zc!=0.0:
-                warning=True
-            if vxc!=0.0 or vyc!=0.0 or vzc!=0.0:
-                warning=True
-
-            if warning:
-                print('Centre is not at origin')
-
-            return xc,yc,zc,vxc,vyc,vzc
-
-        elif self.origin=='cluster':
-            self.xc, self.yc, self.zc = xc,yc,zc
[…]
