### 2023-05-15 #4483 @webbjj
Thanks @lwang-astro  and @rieder for a thorough and very helpful review! I very much appreciate your time.

@lwang-astro - once the JOSS version is done and out there, I will switch my focus to adding PeTar as we discussed awhile ago. The only time consuming part will be allowing for a few more unit options and output types.

@rieder - I addressed all the bugs in your notebook. Thanks for that. One issue was that cluster and direct_cluster had their centres found in different ways, so that is why some calculated values were off. A second was that giving ctype='amuse' and units='pckms' tripped up the code, since it didn't know whether to keep the amuse units or not. I have kept it as is, so if you give units='pckms' it will drop the amuse units and put everything in pckms. However if units is None or units='amuse' then the amuse units are kept. Not sure how this will all work with astropy either. Will probably need to add units='astropy'.

@dfm - I merged your edits to the manuscript. Will tackle the rest of the checklist soon. Thanks for your help.
