+++ tests/test_amuse_orbits.py
@@ -11,7 +11,7 @@ except:
 
 try:
 	import amuse.units.units as u
-    from amuse.units.quantities import ScalarQuantity,VectorQuantity
+	from amuse.units.quantities import ScalarQuantity,VectorQuantity
 
 	noamuse=False
 except:
