+++ clustertools/__init__.py
@@ -1,4 +1,4 @@
-__version__= "1.0.11"
+__version__= "1.1.0"
 from .analysis import *
 from .cluster import *
 from .io import *
