+++ tests/test_amuse_profiles.py
@@ -17,7 +17,7 @@ except:
 
 try:
 	import amuse.units.units as u
-    from amuse.units.quantities import ScalarQuantity,VectorQuantity
+	from amuse.units.quantities import ScalarQuantity,VectorQuantity
 	noamuse=False
 except:
 	noamuse=True
