### 2023-05-17 #4483 @webbjj
Thanks so much @dfm , this is great news. Everything looks good on my end.

I wanted to ask what JOSS' policy is on posting to arXiv (https://arxiv.org/). It would be a great way to share clustertools with the astronomy community. I can post the official JOSS PDF, or I could just post the text. Either way I would provide the link to the JOSS article. Most astronomy journals allow for papers to be posted on the site, just without their official letterhead.

Let me know what works.
