### 2023-05-17 #4483 @dfm
You're welcome to post to ArXiv, but I don't believe that they accept the PDF. I think that this command ☝️ will generate what you need to submit... let's see!
